# openapi_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "bada-openstack-api",
    "version": "0.1.0",
    "description": "bada-openstack-api document"
  },
  "paths": {
...
    "/api/servers/": {
      "get": {
        "tags": [
          "server"
        ],
        "summary": "get list of servers",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "요청 페이지"
          },
          {
            "in": "query",
            "name": "per_page",
            "schema": {
              "type": "integer"
            },
            "description": "한 페이지당 갯수"
          },
          {
            "in": "query",
            "name": "sort_by",
            "schema": {
              "type": "string"
            },
            "description": "정렬 필드명(name, created_at)"
          },
          {
            "in": "query",
            "name": "order_by",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "description": "정렬 조건"
          },
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "description": "[eq/in/not]{value}"
          },
          {
            "in": "query",
            "name": "name",
            "schema": {
              "type": "string"
            },
            "description": "[eq/like]{value}"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ServerResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "server"
        ],
        "summary": "create server",
        "description": "- 서버 이름은 자신이 소유한 다른 서버 이름과 중복해서 생성할 수 없다.\n- 사양은 사전에 제공받은 flavor 중 하나를 선택해서 생성한다.\n- 사전에 제공받은 네트워크와 서브넷을 선택해서 생성한다.\n- 루트볼륨을 생성해야 한다.\n  - 루트 볼륨 타입은 HDD 으로 고정한다.\n  - 이미지는 사전 제공받은 이미지 중 하나를 선택한다.\n- 보안 그룹은 기본 보안 그룹으로 설정한다.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerResponse"
                }
              }
            }
          },
          "400": {
            "description": "필드 조건 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PydanticException"
                }
              }
            }
          },
          "409": {
            "description": "conflict\n1. quota 부족(server / volume/ network)\n2. 이름 중복"
          }
        }
      }
    },
    "/api/servers/{server_id}/": {
      "parameters": [
        {
          "in": "path",
          "name": "server_id",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "server uuid"
        }
      ],
      "get": {
        "tags": [
          "server"
        ],
        "summary": "get specific server",
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "server"
        ],
        "summary": "update specific server info (name & description)",
        "description": "- 서버 이름과 설명을 변경할 수 있다.\n- 서버 이름은 자신이 소유한 다른 서버 이름과 중복할 수 없다",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "server"
        ],
        "summary": "delete specific server",
        "description": "- 서버는 삭제 가능한 상태에서만 삭제할 수 있다.\n- 서버 삭제 시 네트워크 인터페이스는 자동으로 삭제한다.\n- 서버 삭제 시 루트 볼륨은 자동으로 삭제한다.\n- 서버 삭제 시 추가 볼륨은 연결 해제 한다.\n- 서버 삭제 시 플로팅 IP 는 연결 해제 한다.\n- 서버 삭제 시 보안 그룹은 연결 해제 한다",
        "responses": {
          "204": {
            "description": "No-Content"
          }
        }
      }
    },
...
    "/api/servers/{server_id}/vnc/": {
      "parameters": [
        {
          "in": "path",
          "name": "server_id",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "server uuid"
        }
      ],
      "get": {
        "tags": [
          "server"
        ],
        "summary": "get vnc connection url",
        "description": "vnc 접속 url을 제공한다",
        "responses": {
          "200": {
            "description": "return vnc url",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VncUrlResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/servers/{server_id}/volumes/": {
      "parameters": [
        {
          "in": "path",
          "name": "server_id",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "server uuid"
        }
      ],
      "patch": {
        "tags": [
          "server"
        ],
        "summary": "attach/detach volume",
        "description": "생성된 volume을 서버에 연결(is_root=False인 볼륨)하거나 해제한다",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerUpdateVolumeRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/volumes/": {
      "get": {
        "tags": [
          "volume"
        ],
        "summary": "get list of volumes",
        "description": "조건에 해당하는 volume list를 반환한다",
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VolumeResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "volume"
        ],
        "summary": "create volume",
        "description": "create volume",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VolumeCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolumeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/volumes/{volume_id}/": {
      "parameters": [
        {
          "in": "path",
          "name": "volume_id",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "volume uuid"
        }
      ],
      "get": {
        "tags": [
          "volume"
        ],
        "summary": "get specific volume",
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolumeResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "volume"
        ],
        "summary": "update specific volume's info(name, description)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VolumeUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolumeResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "volume"
        ],
        "summary": "delete specific volume",
        "description": "볼륨은 삭제 가능한 상태에서만 삭제할 수 있다.\n\n서버에 연결한 볼륨은 삭제할 수 없다.",
        "responses": {
          "204": {
            "description": "No-content"
          }
        }
      }
    },
    "/api/volumes/{volume_id}/size/": {
      "parameters": [
        {
          "in": "path",
          "name": "volume_id",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "volume uuid"
        }
      ],
      "patch": {
        "tags": [
          "volume"
        ],
        "summary": "update specific volume's size",
        "description": "**need polling**\n\n해당 volume의 크기를 변경(상향만 가능)한다",
        "responses": {
          "202": {
            "description": "accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolumeResponse"
                }
              }
            }
          },
          "400": {
            "description": "field error"
          },
          "409": {
            "description": "하향 불가, 상태 불가"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PydanticException": {
        "properties": {
          "error_type": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "detail": {
            "properties": {
              "input": {
                "properties": {
                  "field_name": {
                    "type": "string"
                  }
                },
                "type": "object"
              },
              "loc": {
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    }
                  ]
                },
                "type": "array"
              }
            },
            "type": "object"
          }
        },
        "type": "object",
        "title": "PydanticException",
        "description": "필드 조건 오류"
      },
...
      "ServerCreateRequest": {
        "title": "ServerCreateRequest",
        "description": "서버 생성시 입력 dto",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "flavor_id": {
            "type": "string"
          },
          "volume": {
            "description": "생성할 루트 볼륨 정보",
            "properties": {
              "size": {
                "type": "integer"
              },
              "name": {
                "type": "string"
              },
              "image_id": {
                "type": "string"
              }
            },
            "required": [
              "size",
              "name",
              "image_id"
            ]
          }
        },
        "required": [
          "name",
          "flavor_id",
          "image_id",
          "volume"
        ]
      },
      "ServerUpdateRequest": {
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "ServerResponse": {
        "properties": {
          "server_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          },
          "flavor": {
            "properties": {
              "flavor_id": {
                "type": "string"
              },
              "cpu": {
                "type": "string"
              },
              "memory": {
                "type": "integer"
              }
            }
          },
          "volumes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VolumeOverallResponse"
            },
            "description": "첫 생성시에는 할당되지 않을 수 있음"
          },
          "floatingip": {
            "$ref": "#/components/schemas/FloatingipOverallResponse"
          },
          "securitygroup": {
            "$ref": "#/components/schemas/SecurityGroupOverallResponse"
          },
          "network_id": {
            "type": "string",
            "description": "private network id"
          },
          "port_id": {
            "type": "string",
            "description": "첫 생성시에는 할당되지 않을 수 있음"
          },
          "fixed_address": {
            "type": "string",
            "description": "첫 생성시에는 할당되지 않을 수 있음"
          },
          "status": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "deleted_at": {
            "type": "string"
          }
        }
      },
      "VncUrlResponse": {
        "properties": {
          "vnc_url": {
            "type": "string"
          }
        }
      },
      "ServerUpdateVolumeRequest": {
        "properties": {
          "volume_id": {
            "type": "string"
          }
        },
        "required": [
          "volume_id"
        ]
      },
      "VolumeCreateRequest": {
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "size": {
            "type": "integer"
          }
        },
        "required": [
          "name",
          "size"
        ]
      },
      "VolumeUpdateRequest": {
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "VolumeResponse": {
        "properties": {
          "volume_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "size": {
            "type": "integer"
          },
          "is_root": {
            "type": "boolean"
          },
          "status": {
            "type": "string"
          },
          "server": {
            "$ref": "#/components/schemas/ServerOverallResponse"
          }
        }
      },
...
      "FloatingipOverallResponse": {
        "description": "연관된 floating ip 보여줄때",
        "properties": {
          "floatingip_id": {
            "type": "string"
          },
          "ip_address": {
            "type": "string"
          }
        }
      },
      "VolumeOverallResponse": {
        "description": "연관된 볼륨 보여줄 때",
        "properties": {
          "volume_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "volume_type": {
            "type": "string",
            "default": "HDD"
          },
          "size": {
            "type": "integer"
          },
          "is_root_volume": {
            "type": "boolean"
          },
          "image": {
            "description": "volume source가 image인 경우",
            "properties": {
              "image_id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            }
          }
        }
      },
      "SecurityGroupOverallResponse": {
        "description": "연관된 보안 그룹 보여줄 때",
        "properties": {
          "securitygroup_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "ServerOverallResponse": {
        "description": "연관된 서버 보여줄 때",
        "properties": {
          "server_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      }
    }
  }
}